fix: sed -i "" not supported on linux
Note that I've tried to use a variable to hold the quotes and it
resulted in producing a backup file with the quotes as a suffix (i.e. ngink.conf.pre"")
It looks like the quotes in the variable are being escaped through the interpolation.
I've tried all sorts of combination and decided to create the
backup file and then remove it.