Fix SVG from bento-web served as text-plain
This caused a bug where SVGs could not be displayed correctly in the
browser (missing image) even if the data was fetched correctly.
The mime-type was text-plain.
It appears that the gateway acts as a reverse proxy for the
nginx server that is located inside bentov2-web.
The nginx config for that server has been modified recently to accomodate
changes made to the node docker container.
The `http` context has been completely overriden, while
previously it was only the `server` directive
in that context that had been touched.
Adding `include mime.types;` directive restores the proper handling of files.