fix add localhost to ALLOWED_HOSTS in production mode
In production mode some calls received from localhost:8000 are rejected
by Django.
There was a sed invocation in the Dockerfile to add localhost:8000, but
it did not work due to 2 issues:
[CHORD_HOST or "localhost:8000"] was always returning only [CHORD_HOST],
because its value is not falsy so the left hand side of the `or`
expression is never evaluated.
Second issue is that `localhost:8000` does not match the host name `localhost`
it is compared to when a call from `localhost:8000` is being made.