fix: latest VEP breaks outdated vcf2maf script
vcf2maf executes ensembl_vep with a long list of arguments.
One of this is deprecated in v107.0, which results in an execution error
when vcf2maf is called.
From https://github.com/mskcc/vcf2maf/issues/322 it
appears that a simple fix exists, but it hasn't been merged
in the last 2 months.
v105.0  introduced new effects predictions which cause warnings in
cBioportal when imported.